# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.CreateRatingRequest:
    properties:
      comment:
        maxLength: 500
        type: string
      is_public:
        type: boolean
      order_id:
        type: integer
      score:
        maximum: 5
        minimum: 1
        type: number
      staff_id:
        type: integer
    required:
    - order_id
    - score
    - staff_id
    type: object
  model.CreateStaffRequest:
    properties:
      id_card:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - id_card
    - name
    - phone
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 滴滴打人服务 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 滴滴打人 API
  version: "1.0"
paths:
...
  /security/rating:
    get:
      consumes:
      - application/json
      description: 获取安保人员的评价列表
      parameters:
      - description: 安保人员ID
        in: query
        name: staff_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取评价列表
      tags:
      - 安保人员
    post:
      consumes:
      - application/json
      description: 为安保人员创建评价
      parameters:
      - description: 评价信息
        in: body
        name: rating
        required: true
        schema:
          $ref: '#/definitions/model.CreateRatingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建评价
      tags:
      - 安保人员
  /security/staff:
    get:
      consumes:
      - application/json
      description: 获取安保人员列表，支持分页和筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      - description: 状态
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取安保人员列表
      tags:
      - 安保人员
    post:
      consumes:
      - application/json
      description: 创建新的安保人员
      parameters:
      - description: 安保人员信息
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/model.CreateStaffRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建安保人员
      tags:
      - 安保人员
  /security/staff/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取安保人员详情
      parameters:
      - description: 安保人员ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取安保人员
      tags:
      - 安保人员
  /security/staff/{id}/status:
    put:
      consumes:
      - application/json
      description: 更新安保人员的工作状态
      parameters:
      - description: 安保人员ID
        in: path
        name: id
        required: true
        type: string
      - description: 状态
        in: query
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新安保人员状态
      tags:
      - 安保人员
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"